Fix PASV reply for Firefox.
Section 4.1.2.6 of RFC 1123 (October 1989):
"The format of the 227 reply to a PASV command is not
well standardized.  In particular, an FTP client cannot
assume that the parentheses shown on page 40 of RFC-959
will be present (and in fact, Figure 3 on page 43 omits
them).  Therefore, a User-FTP program that interprets
the PASV reply must scan the reply for the first digit
of the host and port numbers."

How retarded is the fact that the "standard compliant", "lean and mean"
browser requires these parentheses?